Added zsh check

+++ MakeItReset.sh
@@ -1,4 +1,11 @@
 #!/bin/sh
+pat=`echo $SHELL`
+if  [[ $pat == *zsh  ]]  ;
+	then
+		echo "Zsh will be removed, might want to switch to bash: chsh -s /bin/bash" 
+		exit 1
+fi
+echo "Let's Go!!!!!!!!!!!!!!!!!!!!!!!!!!!!!"
 sudo dpkg --clear-selections
 sudo dpkg --set-selections < list.txt
 sudo apt-get autoremove
